+++ src/main/java/ru/yandex/money/common/dbqueue/internal/runner/QueueRunnerInTransaction.java
@@ -1,7 +1,7 @@
 package ru.yandex.money.common.dbqueue.internal.runner;
 
 import ru.yandex.money.common.dbqueue.api.QueueConsumer;
-import ru.yandex.money.common.dbqueue.dao.QueueDao;
+import ru.yandex.money.common.dbqueue.api.QueueShard;
 import ru.yandex.money.common.dbqueue.internal.QueueProcessingStatus;
 import ru.yandex.money.common.dbqueue.settings.ProcessingMode;
 
@@ -19,7 +19,7 @@ import java.util.Objects;
 class QueueRunnerInTransaction implements QueueRunner {
 
     @Nonnull
-    private final QueueDao queueDao;
+    private final QueueShard queueShard;
     private final BaseQueueRunner baseQueueRunner;
 
     /**
@@ -27,18 +27,18 @@ class QueueRunnerInTransaction implements QueueRunner {
      *
      * @param taskPicker    выборщик задачи
      * @param taskProcessor обработчик задачи
-     * @param queueDao      шард на котором обрабатываются задачи
+     * @param queueShard    шард на котором обрабатываются задачи
      */
     QueueRunnerInTransaction(@Nonnull TaskPicker taskPicker, @Nonnull TaskProcessor taskProcessor,
-                             @Nonnull QueueDao queueDao) {
-        this.queueDao = Objects.requireNonNull(queueDao);
+                             @Nonnull QueueShard queueShard) {
+        this.queueShard = Objects.requireNonNull(queueShard);
         baseQueueRunner = new BaseQueueRunner(taskPicker, taskProcessor, Runnable::run);
     }
 
     @Override
     @Nonnull
     public QueueProcessingStatus runQueue(@Nonnull QueueConsumer queueConsumer) {
-        return queueDao.getTransactionTemplate().execute(status ->
+        return queueShard.getTransactionTemplate().execute(status ->
                 baseQueueRunner.runQueue(queueConsumer));
     }
 }