+++ src/main/java/ru/yandex/money/common/dbqueue/internal/runner/TaskProcessor.java
@@ -1,11 +1,11 @@
 package ru.yandex.money.common.dbqueue.internal.runner;
 
 import ru.yandex.money.common.dbqueue.api.QueueConsumer;
+import ru.yandex.money.common.dbqueue.api.QueueShard;
 import ru.yandex.money.common.dbqueue.api.Task;
 import ru.yandex.money.common.dbqueue.api.TaskExecutionResult;
 import ru.yandex.money.common.dbqueue.api.TaskLifecycleListener;
 import ru.yandex.money.common.dbqueue.api.TaskRecord;
-import ru.yandex.money.common.dbqueue.dao.QueueDao;
 import ru.yandex.money.common.dbqueue.internal.MillisTimeProvider;
 
 import javax.annotation.Nonnull;
@@ -22,7 +22,7 @@ import static java.util.Objects.requireNonNull;
 class TaskProcessor {
 
     @Nonnull
-    private final QueueDao queueDao;
+    private final QueueShard queueShard;
     @Nonnull
     private final TaskLifecycleListener taskLifecycleListener;
     @Nonnull
@@ -33,16 +33,16 @@ class TaskProcessor {
     /**
      * Конструктор
      *
-     * @param queueDao              шард на котором происходит выполнение задачи
+     * @param queueShard            шард на котором происходит выполнение задачи
      * @param taskLifecycleListener слушатель жизненного цикла задачи в очереди
      * @param millisTimeProvider    поставщик текущего времени
      * @param taskResultHandler     обработчик результата выполнения задачи
      */
-    TaskProcessor(@Nonnull QueueDao queueDao,
+    TaskProcessor(@Nonnull QueueShard queueShard,
                   @Nonnull TaskLifecycleListener taskLifecycleListener,
                   @Nonnull MillisTimeProvider millisTimeProvider,
                   @Nonnull TaskResultHandler taskResultHandler) {
-        this.queueDao = requireNonNull(queueDao);
+        this.queueShard = requireNonNull(queueShard);
         this.taskLifecycleListener = requireNonNull(taskLifecycleListener);
         this.millisTimeProvider = requireNonNull(millisTimeProvider);
         this.taskResultHandler = requireNonNull(taskResultHandler);
@@ -58,22 +58,22 @@ class TaskProcessor {
         requireNonNull(queueConsumer);
         requireNonNull(taskRecord);
         try {
-            taskLifecycleListener.started(queueDao.getShardId(), queueConsumer.getQueueConfig().getLocation(),
+            taskLifecycleListener.started(queueShard.getShardId(), queueConsumer.getQueueConfig().getLocation(),
                     taskRecord);
             long processTaskStarted = millisTimeProvider.getMillis();
             Object payload = queueConsumer.getPayloadTransformer().toObject(taskRecord.getPayload());
-            Task task = new Task(queueDao.getShardId(), payload, taskRecord.getAttemptsCount(),
+            Task task = new Task(queueShard.getShardId(), payload, taskRecord.getAttemptsCount(),
                     taskRecord.getCreateDate(), taskRecord.getCorrelationId(), taskRecord.getActor());
             TaskExecutionResult executionResult = queueConsumer.execute(task);
-            taskLifecycleListener.executed(queueDao.getShardId(), queueConsumer.getQueueConfig().getLocation(),
+            taskLifecycleListener.executed(queueShard.getShardId(), queueConsumer.getQueueConfig().getLocation(),
                     taskRecord,
                     executionResult, millisTimeProvider.getMillis() - processTaskStarted);
             taskResultHandler.handleResult(taskRecord, executionResult);
         } catch (Exception exc) {
-            taskLifecycleListener.crashed(queueDao.getShardId(), queueConsumer.getQueueConfig().getLocation(),
+            taskLifecycleListener.crashed(queueShard.getShardId(), queueConsumer.getQueueConfig().getLocation(),
                     taskRecord, exc);
         } finally {
-            taskLifecycleListener.finished(queueDao.getShardId(), queueConsumer.getQueueConfig().getLocation(),
+            taskLifecycleListener.finished(queueShard.getShardId(), queueConsumer.getQueueConfig().getLocation(),
                     taskRecord);
         }
     }